# APIGOLANGMAP/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/position": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria uma localizacao de um utilizador em especifico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Position"
                ],
                "summary": "Adicionar uma localizaçao",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add Location",
                        "name": "evaluation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.PositionAddSwagger"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Position"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/position/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exibe a lista da última localização do utilizador",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Position"
                ],
                "summary": "Obter a última localização do utilizador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Position"
                        }
                    },
                    "400": {
                        "description": "User Token Malformed"
                    },
                    "404": {
                        "description": "User Not found"
                    }
                }
            }
        },
        "/position/filter": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exibe a lista de localizações dos utilizadores",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Position"
                ],
                "summary": "Obtem todas as localizações dos utilizadores com filtros",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add Location",
                        "name": "localization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.UserLocationWithFilters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Position"
                            }
                        }
                    },
                    "400": {
                        "description": "User Token Malformed"
                    },
                    "404": {
                        "description": "Location Not found"
                    }
                }
            }
        },
        "/position/history": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exibe a lista de todas as localizações do utilizador",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Position"
                ],
                "summary": "Obtem todas as localizações do utilizador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "get locations history",
                        "name": "localization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.UserLocationHistory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Position"
                            }
                        }
                    },
                    "400": {
                        "description": "User Token Malformed"
                    },
                    "404": {
                        "description": "User Not found"
                    }
                }
            }
        },
        "/position/history/user": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtem todas as localizações do utilizador",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Position"
                ],
                "summary": "Obtem todas as localizações do utilizador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add Location",
                        "name": "localization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.UserLocationHistoryUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Position"
                            }
                        }
                    },
                    "400": {
                        "description": "Check Syntax, Dates Malformed"
                    },
                    "404": {
                        "description": "UserID Not found"
                    }
                }
            }
        },
        "/position/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exclui uma localização selecionada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Position"
                ],
                "summary": "Exclui uma localização",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Position ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete succeeded!"
                    },
                    "404": {
                        "description": "None found!"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Position": {
            "type": "object",
            "required": [
                "Latitude",
                "Longitude"
            ],
            "properties": {
                "Latitude": {
                    "type": "number"
                },
                "Longitude": {
                    "type": "number"
                },
                "UserId": {
                    "type": "integer"
                }
            }
        },
...
        "swaggermodel.PositionAddSwagger": {
            "type": "object",
            "required": [
                "Latitude",
                "Longitude"
            ],
            "properties": {
                "Latitude": {
                    "type": "number",
                    "example": 39.557191
                },
                "Longitude": {
                    "type": "number",
                    "example": -7.8536599
                }
            }
        },
...
        "swaggermodel.UserLocationHistory": {
            "type": "object",
            "required": [
                "end",
                "start"
            ],
            "properties": {
                "end": {
                    "type": "string",
                    "example": "2022-07-01"
                },
                "start": {
                    "type": "string",
                    "example": "2022-01-01"
                }
            }
        },
        "swaggermodel.UserLocationHistoryUser": {
            "type": "object",
            "required": [
                "end",
                "start",
                "userID"
            ],
            "properties": {
                "end": {
                    "type": "string",
                    "example": "2022-07-01"
                },
                "start": {
                    "type": "string",
                    "example": "2022-01-01"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "swaggermodel.UserLocationWithFilters": {
            "type": "object",
            "required": [
                "UserId",
                "end",
                "start"
            ],
            "properties": {
                "UserId": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "end": {
                    "type": "string",
                    "example": "2022-07-01"
                },
                "start": {
                    "type": "string",
                    "example": "2022-01-01"
                }
            }
        },
...
    }
}